Reorganize conditional tasks
