+++ Executor/Tasks/Flow/__init__.py
@@ -1,2 +1,3 @@
 from .parallel import Parallel
-from .sequences import Sequence, Repeat, While, Select
+from .sequences import Sequence, Repeat
+from .conditionals import While, Select
